techdoc9.html
This commit provides a correction to the previous commit, techdoc8.html. This commit contains an updated comment which aknowledges the completion of User Stories #11 (The header element in the navbar must come before any link (a) elements in the navbar), #12 (Each element with the class of nav-link should contain text that corresponds to the header text within each section), and #13 (When I click on a navbar element, the page should navigate to the corresponding section of the main-doc element).